+++ reinvent/runmodes/create_adapter.py
@@ -1,6 +1,7 @@
 """Create a model adapter from a Torch pickle file"""
 
 __all__ = ["create_adapter"]
+import pprint
 from typing import Tuple
 import logging
 
@@ -22,10 +23,27 @@ def create_adapter(dict_filename: str, mode: str, device: torch.device) -> Tuple
     """
 
     # FIXME: check if map_location is needed for CPU as in Reinvent
-    save_dict = torch.load(dict_filename, map_location=device)
+    #save_dict = torch.load(dict_filename, map_location=device)
+    save_dict = torch.load(dict_filename, map_location="cpu")
 
-    if "model_type" in dir(save_dict):
-        model_type = save_dict._model_type
+    if "metadata" in save_dict:
+        metadata: models.ModelMetaData = save_dict["metadata"]
+
+        if not metadata.hash_id:
+            logger.warning(f"{dict_filename} does not contain a hash ID")
+        else:
+            valid = models.check_valid_hash(save_dict)
+            pp = pprint.PrettyPrinter(indent=2)
+
+            if valid:
+                logger.info(f"{dict_filename} has valid hash:\n{pp.pformat(metadata.as_dict())}")
+            else:
+                logger.error(f"{dict_filename} has invalid hash:\n{pp.pformat(metadata.as_dict())}")
+    else:
+        logger.warning(f"{dict_filename} does not contain metadata")
+
+    if "model_type" in save_dict:
+        model_type = save_dict["model_type"]
     else:  # heuristics
         # FIXME: ugly if
         if "network" in save_dict: